OO-219: refactoring, extracting report generating to defaultrouteservice
